Split provider status and config
