+++ pkg/apis/awsprovider/group.go
@@ -14,5 +14,5 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package awsproviderconfig contains awsprovider API versions
-package awsproviderconfig
+// Package awsprovider contains awsprovider API versions
+package awsprovider
